clean code：混用有符号数和无符号数
# message auto-generated for no-merge-commit merge:
merge clean-code into master

clean code：混用有符号数和无符号数

Created-by: msprofDeveloper

Match-id-643468a5badbedfd0fb0a4b80a8501a677563a96

+++ analysis/csrc/application/credential/id_pool.cpp
@@ -48,7 +48,7 @@ uint64_t IdPool::GetUint64Id(const std::string& key)
 }
 
 // 获取string格式map的key对应value(32位)，如果不存在则添加(key, value)并返回value
-uint64_t IdPool::GetUint32Id(const std::string& key)
+uint32_t IdPool::GetUint32Id(const std::string& key)
 {
     std::lock_guard<std::mutex> lock(uint32Mutex_);
     if (uint32Ids_.find(key) == uint32Ids_.end()) {
